Reorganize Sidebar components
