+++ src/lib/components/sidebar/SidebarItem.svelte
